[dev][usb-bus] Add support for overriding default configuration based on VID/PID
By default, we create devices for interfaces on the USB device's first configuration
This change adds a hard coded list of VID/PID pairs to specify using a different
configuration instead. This is just a short term fix until we have a better solution
for dealing with devices with multiple configurations.

Also switch to configuration 2 for a particular Realtek ethernet dongle so we can
use its CDC interface instead of the vendor proprietary interface.

Change-Id: Ife2025a7ac91e39284bc36d678fc2ec64b9893cc