[libtftp] Use opcode prefixes
Use the (previously-unused) upper byte of the opcode to store an index
value that we increment on evidence of transmission failure (e.g.,
unexpected DATA packets, timeouts, or unexpected ACKs). This helps
us to work around an issue with the asix 88179 chipset where invalid
IPv6 checksums are calculated. Adding some variation on retransmission
allows a valid packet to be generated.

Change-Id: If06211707e2861577643fd8467b0edda2e9e9d02