Enable fences in Shadertoy service.
Last time I tried this it crashed; something has improved in the driver.

However, all is not perfect: in debug mode Vulkan validation is enabled,
and there is much log-spam about signaling semaphores that have already
been signaled, and waiting on semaphores that will not be signaled.  It
seems that the validation layers do not work properly with the semaphore
import/export extension.

Bug: MZ-240
Change-Id: I1b2ca4fdc30b85947f371307c6267307af11395e