apps/media: add tasks to the in-proc framework
Tasks are a new framework concept that allows code to be run against
nodes in the graph exclusive of any updates or other tasks that might
involve those nodes. A new Graph::PostTask method runs a function
once a list of nodes has been acquired. Tasks posted in this way can
run in any order if their set of nodes don't intersect. If two tasks
are posted referencing the same node, those tasks will be run in the
order they were posted. A task may be posted on an arbitrary thread
and will run on an arbitrary thread, possibly synchronous with the
PostTask call.

Change-Id: Iae1f72a6d7dee2a0c247dcdb56adac3ceddf47af