Make daemon.GetUpdates a synchronous call
GetUpdates now blocks until updates are retrieved and stored
locally. Concurrent calls to GetUpdates are allowed. The Daemon
keeps a map that tracks the in-flight calls go get updates.
If call A requests package cde456 and call B requests cde456 and
arrives while call A is fetching cde456, call B subscribes to the
result of call A. If B also requests 123abc, B starts a routine to
listen to its own result and fetches 123abc. B then waits for the
routines waiting for the results from A and B before finally
returning.

The next update will rework source_monitor.

Change-Id: I587289ad4ee7da917b44b604340655369a1ea30f