[hypervisor][block] Fix handling of multiple queue descriptors.
Stateful context structures are passed along during virt queue
processing. In the case there were multiple descriptors in the
available ring, this state was carried on between requests.

With the block device, this was causing us to accumulate incorrect
offsets into the file since we were reusing offset values caculated
for a previous request.

This change causes the virtio queue method to only process a
single descriptor from the available ring at a time so the caller can
reset any state used by the request function.

Change-Id: Iebcdc917b8acceffa4f83b27ac0bb74c64aaf211