[kernel][magenta] fix port_cancel() race
port_cancel() was failing with MX_ERR_NOT_FOUND
for MX_WAIT_ASYNC_ONCE waits because there was a
logical race between PortObserver::OnRemoved() and
PortDispatcher::CancelQueued() because a thread can
get in between the two calls and it is visible
when using PortPacket::observer member which was
doing double duty as "packet is a signal" and
"packet needs to be deleted".

The fix is to keep the observer member just
as handling the delete, so we add a handle member
to PortPacket so we stop relying on the observer
for the other case.

Note that heavier locking could also fix this
problem but as evidenced by this fix the problem
can be solved by using copies of already inmutable
data.

A new test case added, which probably needs
a bigger number of retries for doing manual
validation but kept low so automated tests don't
take too long.

Future historians should also look at PS1 for
a worse fix that makes clear the problem.

Change-Id: I8b1c4eaa136f2b18273717c5134d7eb648494e4b