[sysgen] Handle "optional" attribute on secondary return values
The "optional" attribute on a secondary return value says that in
C-like APIs where secondary return values are returned via "out"
pointer arguments, it's OK if the caller passes a null pointer so as
to ignore that return value.  For all other "out" pointer arguments
generated for secondary return values, the C function declaration
uses the "nonnull" attribute to enable warnings for calls that the
compiler can tell at compile time will pass a null pointer.

Change-Id: Ia56dafd60fc98defc419ac262cd40033ef408c05