[system] [core] getenv_bool for cmdline bools
The core kernel behavior for parsing boolean flags is supported by a
cmdline_get_bool. devmgr maps command line arguments to environment
variables. getenv_bool implements the same value semantics as
cmdline_get_bool. The same semantics are also applied in virtcon as a
special case, as it is the only place in core, outside of devmgr, that
inherits boolean options.

Change-Id: Icff00d9a2597ab8fa482f30002120717ce61748b