apps/media: make FactoryServiceBase into a template
FactoryServiceBase is a base class for services that are factories
for multiple fidl components. media_service, net_media_service and
flog_service implementations all use this class. FactoryServiceBase
defines the relationship between the factory and its 'products'.

Previously, a product's reference to the factory had the type
FactoryServiceBase. This prevented characteristics particular to
the factory subclass from being exposed to the products. This CL
makes FactoryServiceBase into a template, so we have:

  class MyFactory : public FactoryServiceBase<MyFactory>

instead of

  class MyFactory : public FactoryServiceBase

The products' 'owner' (factory) reference is, in this example, of
type MyFactory rather than FactoryServiceBase, so products can access
methods specific to MyFactory.

Change-Id: I5b093931c1aff56445c7e91e8844a09ade47c1bd