[utest] Remove empty-string msg args from EXPECT_*() calls in C++
The msg argument to EXPECT_*() and ASSERT_*() is optional in C++ now,
so we can drop this argument when it's an empty string.

In the earlier conversions, I only converted ASSERT_*() and forgot to
convert EXPECT_*() too.

This change also converts a few ASSERT_*() instances that appeared
since the earlier changes.

MG-905

Change-Id: I034c6d6a6b614ab0b315dcd409435c9fe9e27e6c