Track if a BSS is an RSN.
When sending an assocreq we should include the RSNE if we are
associating with an RSN. We will only support WPA2-PSK for now and don't
have to include more RSN information in the BSSDescription for now.
However, ultimately we should also keep track of supported ciphers and
allow cipher selection from further up the stack.

This CL will be used to block incoming data frames after associating
with an RSN and forward EAPOL frames to the SME.

We have these resulting follow-up CLs:
- send supported cipher capabilities via RSNE in assocreq
- block data frames on controlled, blocked ports
- forward EAPOL frames for processing (4 Way Handshake for now)

Change-Id: Iea2960e59989053d700d9dabf05e5f4f10398b2c