apps/media: add 'hold frame' to framework flush
We recently added the 'hold frame' concept to flushing at the fidl
level. This is an indication for video renderers fo whether they
should hold and display the newest frame (true) or eject all frames
and display nothing (false).

This CL adds a hold_frame parameter to flushing in the in-proc
framework. This will be needed when we have an ActiveSink video
renderer. The current video renderer doesn't use the framework.

Change-Id: I41a5e0939b34d1e89c0a6642ca495ab84696626d