apps/media: fix media_player_flutter ui to work around MZ-252
In order for UI to float over video, we need to wrap the UI in
PhysicalModel widgets, which are maybe-rounded rectangles that cannot
be transparent (yet). This CL adds the PhysicalModel widgets and
otherwise tweaks the UI so it looks reasonable given these
constraints.

For some reason, the PhysicalModel thing is required for the X in
the upper right of the chooser UI, even though there's no child
view involved.

The icon in the upper right of the 'playback screen' in the
example app has been wrapped in an Offstage widget so it
disappears when the rest of the player UI disappears.

Change-Id: I408b916f700e815f3b261c80f18dce19639ea6f8