[hypervisor][pci] Move pci_device_t storage out of pci_bus_t.
Move the device storage out of the bus to allow make it easier for
different compilation units to define their own device device
implementations.

Initialization logic can connect a device to the bus using the
'pci_device_connect' method. Any PCI device slots that have not
been connected will behave as if the device is not implemented.

Change-Id: I68072f4306bd2a45660d525067217c9367f73562