[dev][usb] Fix usb_set_interface() for devices that span multiple interfaces
For certain USB classes (like USB audio and CDC ethernet) we combine
more than one USB interfaces into a devmgr device.
In the case of USB audio, only one of these interfaces have endpoints,
but in the CDC ethernet case each interface has endpoints.

In this change we fix the implementation usb_set_interface() to only
change the state of the endpoints for the interface that is getting
an alternate setting so the endpoints of the other interfaces are not affected.

Change-Id: I18e586e59820298a02134543e2cd3ea4b6473e34