[mxtl] Add replace methods to containers.
Add replace_if to our sequence containers and insert_or_replace to
our associative containers.

replace_if(...) will find the first element in a sequence container (a
singly or doubly linked list) which satisfies a supplied predicate and
replace it using the reference to the supplied by the user.  Then
return the container's reference to the replaced element.

Behavior varies if no element satisfies the supplied predicate
depending on whether the user chose to pass a const reference to a
pointer, or attempted to move a pointer into the container.  If the
user chose to move a pointer, the replace operation has failed and the
pointer will be returned to the user instead of the pointer to a
replaced element.  This allows the user to attempt to replace a
unique_ptr in the container, and fail without accidentally losing their
unique_ptr.  In the case of passing a const reference to a pointer,
the replace operation has failed and nullptr is returned.

insert_or_replace(...) is a similar method for associative containers.
If will search the container for an element with the same key as the
element provided by the user.  If it finds one, it will replace the
element in the container with the provided element and return a
reference to the replaced element.  If no element shares a key with
the element supplied, the element will simply be added to the
collection.

MG-1058 #done

Change-Id: Id62c3165e5ebc3a1b9809da9ce879571cd7b9dfe