Introduce netiface type and fix netif sorting logic
If the netiface has a route for a specific address (e.g. loop back
address), the element should sort before the one with a route for
Any address (e.g. 0.0.0.0).

Also when testing whether they have a gateway or an IP address,
immediately return true or false if only one of them has it.

Change-Id: Ie3691237de0530acd511fe2e19c45fc04b4e4f01