Re-land "Prefer routes to interfaces with a gateway."
Prefer routes to interfaces with a gateway.

NET-130 # done

This original commit was reverted because it was breaking loopback
(because loopback doesn't have a gateway, the route was demoted
and the loopback packets were sent to the other interface with a
gateway incorrectly).

We re-land this commit together with the next commit that fixes
the sort logic (Change-Id: Ie3691237de0530acd511fe2e19c45fc04b4e4f01).

Change-Id: I7d330d47b5031f10f4be9e6c9163a8ad8eb64192