[context] Use metadata-based querying in ContextEngine.
The ContextEngine now stores values in a multi-parent hierarchy keyed by
an internal value ID, instead of storing each value based on a topic
string key. Unlike the topic string, value IDs are created by the
context engine, not the client. Values also now are represented by a
type, content and metadata.

The |type| indicates the value's node type, such as STORY, MODULE,
ENTITY, etc. It is easy to add new node types. The |content| is
currently a bytes that represent JSON string, but will be replaced with
an Entity reference. The |metadata| is a structure that includes
information about the value, such as what Entity type it is, what Story
it is part of, etc.

Values in the graph inherit metadata from their parents. This allows an
Entity value to be a child (or grand-child) of a Story value (or a Link
value, in the future), and adopt its Story-specific metadata (its ID, is
it focused?  visible? etc).

Although the engine itself does not enforce a specific structure to the
graph (ie, what value types can be children of other value types),
conventional use makes Module values children of Story values, etc.

Queries are now performed by specifying a value type and a subset of the
possible metadata fields that the value may contain. This allows for
queries like the following:

* "What Entities of type 'hotel' are visible in the currently focused
story?"
* "What are all the Modules running for the current user?"
* "What Entities are referenced in the Link with name 'foo'?"

The following work is necessary but can be done in follow-up CLs:

* Add parent value ID handling context_writer_impl
* Add Agent value type
* Add Link value type

MW-137 #comment

Change-Id: I56e0754ff974efd3ec314f741af9ac4c65a0638a