apps/media: get video playback working again
This CL partially address MTWN-24. The multiproc rework of the
in-proc framework has introduced some regressions that were not
immediately apparent. This CL makes the components in question
single-threaded again and fixes a race in the stage used for the
demux. There are still some issues to resolve, but this gets
playback working again.

Change-Id: I29e91fdcf0b255adbfcdcad8fd4b9996a6587387