[bootloader] Load self-contained mxboot.bin second stage if it exists
If mxboot.bin exists and it consists of a kernel bootdata item
followed by additional items, the additional items are treated
as the "ramdisk", and that kernel and ramdisk combo are started,
instead of following the normal boot path.

Creating a magentaboot image:
% cd build-magenta-pc-x86-64
% echo "netsvc.netboot=true" > cmdline
% ./tools/mkbootfs -o mxboot.bin magenta.bin -C cmdline bootdata.bin

The mxboot.bin is ~2MB smaller if test/... are omitted and could
be made smaller yet by removing various misc commandline tools, etc.

Change-Id: I42059c2d065249715356b13febb9bf1f5471f045