[gap] Zero-length Advertising Manufacturer Data
Advertising Data is allowed to have a Manufacturer Specific Data field
or a Service Data field which has zero data length.

Make the DynamicByteBuffer constructor with a zero length argument
equivalent to the one with no argument.

Add accessors to determine which Manufacturer IDs and UUIDs have service
data set, to disambiguate from one that is set but zero length.

Change-Id: I08c2bbdf5fa00442666aed50d57b33fb611f749c