Add configs/variants for sanitizers (asan, ubsan, sancov)
This lights up the new `select_variant` build argument mechanism.
e.g., this works:
  select_variant = [
    {
      host = true
      name = ["protoc"]
      variant = "asan-sancov"
    },
    {
      host = true
      dir = ["//third_party/yasm"]
      variant = "asan_no_detect_leaks"
    },
    {
      host = true
      name = ["dart"]
      variant = "asan_no_detect_leaks"
    },
    {
      host = true
      variant = "asan"
    },
  ]
to build all the host tools with ASan, and protoc also with coverage,
but the yasm tools and dart with a variant of ASan that disables leak
checking by default, because these tools as used in the build have
some leaks.

Change-Id: I0b53475f286acb9693c9a5fbc3162a38cfbbf28b