Allow specification of options for publishing
Allow the FAR staging directory to be specified as well as the
key for signing the FAR. If no key is specified one is generated.

Allow the update repo to be specified, this will be the
destination of the meta FAR and content blobs, suitable for
serving to clients.

Command takes the form "fpublish [options...] [pkg]".

Change-Id: I0f5d495512e28c486114a3e03f6aec0d90e0ff26