[musl][threads] Remove races in thread creation
After starting the new thread, it may immediately exit and clean up
the thread structure. It's not safe to access the struct after that
call.

Also, write the thrd_t pointer out before starting. Newly C11 threads
synchronize with the end of the thread creation function, so the side
effects must be visible prior to that.

Change-Id: I76ca2563507106dfc2ea2384002606306654ecf5