apps/media: fix dcheck in ActiveMultistreamSourceStage
This CL fixes a problem in ActiveMultistreamSourceStage that was
causing a DCHECK. The class uses a boolean field
(packet_request_outstanding_) to indicate whether the source has
been asked for a packet. This was DCHECKED when a packet was
supplied to make sure the source is behaving properly. When a
flush occurs, packet_request_outstanding_ is cleared, so we
were DCHECKing when the packet was actually delivered. The
correct response is to simply discard the packet.

Change-Id: I55b7e1ed76753da91bd03884d4aa2f6570e1308b