[guest][balloon] Add simple balloon management thread.
Add an optional feature to the guest that will poll virtio-balloon
stats at a given interval and will reap any unused memory above a fixed
threshold from the guest.

Also enable the 'deflate on OOM' feature to allow the guest to
reclaim memory from the balloon without notifying the device if the
memory required for system stability.

Change-Id: Ic64edc65d9df1b3e1221ba1455b1052dd21e26ae