[musl] Fix the race again in thrd_create
The indirection added to fix the restrict violation reintroduces the
race between the start of the new thread and the side effects
(including writing the out parameter) of the calling thread. It seems
better for the moment to fix thread self detaching, and then invert
the flow of control between thrd_create and pthread_create.

Revert "[musl][threads] Fix a restrict violation between thrd_create and pthread_create"

This reverts commit 0f118a2f01a4f096d0d863e28a5f78cbc9bf8ef8.

Change-Id: Id9bccfa85c2ced24c6916ea8f9de4b7b31ceccf6