[async] Split libasync into three parts.
Previously we built libasync as a shared library but that's not
quite correct.  The client and message loop implementations should
be static libraries.  It's only the default dispatcher functions
that need to be shared.

So we split the library into three parts: libasync.a, libasync-loop.a,
and libasync-default.so.

This also saves the linker some trouble since many clients of libasync
only need libasync.a and not the other parts.

Also improved the docs in a few areas to ensure we cover the C++ API
more thoroughly.

Change-Id: I98702dab4dfc21cbce87b8acddfc046a449410e6