[gap] Introduce LowEnergyConnectionManager
LowEnergyConnectionManager is responsible for managing LE logical
links. It provides a class API for many clients to request and share
ownership of a connection to a peripheral and for listeners to be
notified on connection creation.

LowEnergyConnectionManager also updates the L2CAP layer and
RemoteDeviceCache as connections are added and removed.

Currently supports only the Direct Connection Establishment Procedure,
however other procedures (General & Auto) will also be managed by
this class.

Change-Id: Ief4c9a8f5d0466afe9a6c219af6df6283a8b1e19