[ulib][musl] Increase PTHREAD_STACK_MIN to 3072
This constant is used by launchpad to decide how big the initial
stack needs to be (in addition to the bootstrap message size).
So it must be sufficient for all the stack use in the dynamic
linker and libc startup path (not including the bootstrap message
size), up until the point where __libc_start_main switches
stacks.  The actual stack size is always rounded up to whole
pages, so it's worthwhile to keep this size smaller than a page
so that when the bootstrap message is small the initial stack
only needs one page.

Current stack usage is comfortably under 3k, but more than 2k.
So with the old size of 2k, a bootstrap message size (modulo page
size, 4k) that is between around 1500 and around 1900 would be
too large to leave enough stack space but too small to push
launchpad's calculation up to using another page of stack.

MG-1038 #done

Change-Id: Ia1aadc39aa4717f8ff75ade0087eccb74cafb6d2