[hypervisor][virtio] Implement virtio ISR register.
Bits in the Virtio ISR register inform the driver if it's a device
configuration or queue that needs attention. Replace the hard-coded
assumption that all interrupts are are due to queue operations.

This implements the ISR register for queue events while exposing a path
for devices to implement configuration change interrupts.

Change-Id: Iaad3d43d0d4d3f5ac35a1cd8d7997d726fcb8182