Fix shutdown race in reporter thread
This fixes a race condition in the Reporter thread when the TestRunner
connection fails at the same time as the last test runs. Failing the
race looked like this:

1.) Test starts running
2.) Reporter::Start() issues a connect to to the TestRunner service
3.) All tests finish
4.) Reporter::Stop() issues a call to test_runner_->Teardown
5.) TestRunner connection fails, and connection error handler runs
6.) Reporter::test_runner_ is nulled out and the callback to Teardown()
    is dropped on the floor

This patch makes Reporter remember if it is attempting to stop and
posts a quit task if it observes an error on the TaskRunner connection
and is already stopping.

Change-Id: Ifa2a21bdaca2b78f38b80d0f6d4306508c599760