[crashlogger] Bail out of dso list if too many entries detected
Crashlogger reads a value out of a process' property table and then
examines memory, hoping to find a data structure describing the DSOs in
that process. Since the property can point to anywhere it's possible
that the data structure in a faulting process is corrupt or that the
property refers to a region of memory completely unrelated to the DSO
list. This keeps track of how many DSOs appear to be in the data
structure and bails out if the number is too large to avoid looping
infinitely on bad data. The individual data values are already sanity
checked to some degree.

MG-1040 #done

Change-Id: Ibbf29e4fd3e3e3f825862fb6ffa08f803ca4bfce