apps/media: improve Stage/Node relationship
This CL improves the way stages relate to the nodes they host, fixing
some long-standing issues with the in-proc streaming framework.

We now have base classes for Node and Stage in framework/model. These
classes provide some common capabilities to all nodes (including
access to their respective stages) and establishes a model for stages
from the perspective of the node. For now, Stage is empty, but that
will change.

Each node model definition now defines both node abstract class (a
subclass of Node) and the stage as seen from that type of node (a
subclass of Stage). Nodes no longer have SetXCallback methods,
because they can now simply call their stages.

The old Stage class has been renamed StageImpl. It still defines a
stage as the graph and engine see it.

Change-Id: I9c95ca9423b79ebae6553fc2aa24a65c3fade7b1