Remove extraneous thread for test reporting
The test_runner reporting library provides a facility for reporting
test results from GTest suites to a TestRunner service provided in the
environment. To do this, it used to require that everyone using the
reporter spin up an additional thread with its own message loop and
requires that all test results be queued up for transmission. This is
complicated and error-prone (see
https://fuchsia-review.googlesource.com/#/c/test_runner/+/52465/) and
completely unnecessary - our system does not require spinning up an
additional thread in order to send messages.

This rips out the extra thread and uses the synchronous C++ bindings to
report test results from whatever thread is running the test.

This also removes racy initialization of the tracing provider from a
non-main thread.

Change-Id: I26bbddef7d3d47d9d50cb6114d652878d0a4e9d0