[mkbootfs] allow manifest lines to specify groups
A manifest line beginning with {groupname} indicates that that file
belongs to the group "groupname".  By default all groups are included,
but the -g option may be used to restrict group inclusion to a comma
separated list of specific group names.  Filtering may be disabled
again by specifying -g all

Change-Id: I8722642eb99d5a2a1a67a8e49e334ea97df93eab