Map/Scope '/data' directory for some applications to their parent's namespace.
(1) We modify AppClient(Base) to optionally take a path as argument and
map it into the newly started application's namespace at /data.
(2) We modify device_runner which has access to the root namespace to
create a directory strucutre like:
/data/modular
├── ACCOUNT_MANAGER
├── USER_ABC
├── USER_PQR
└── USER_XYZ
(3) ACCOUNT_MANAGER is mapped to /data of Account Manager.
(4) When logging in as a non-guest user, say 'USER_ABC' we map
device_runner's /data/modular/USER_ABC to the user_runner's /data.
(5) For Guest users we map user_runner's /data from device_runner's
/tmp/GUEST_USER_ABC.
(6) The user_runner's /data/LEDGER is mapped to /data for the ledger.
This means that we don't need to explicitly provide repository path to
ledger.
(7) Since we now store guest mode on /tmp, removing a TODO about garbage
collecting guest user data.

Change-Id: Icbdb16a49d848706cc09b94eb3a0ed63377fad6c