apps/media: use task runners in framework
This CL changes the in-proc framework so it no longer uses an update
work list across all stages/nodes and instead uses task runners that
can be specific to a give stage/node pair.

StageImpl now implements a PostTask method that's available to the
stage itself and to the node it hosts. StageImpl::PostTask uses its
own queue and employs an externally-supplied task runner to run tasks
from that queue. The private task queue is used so that, regardless
of the characteristics of the externally-supplied task runner, all
tasks posted for a given stage/node are mutually exclusive. This task
queue will also have the ability to suspend, which will be useful for
dynamic configuration changes.

Ultimately, tasks posted using StageImpl::PostTask are run on a task
runner supplied to the stage. Task runners can be supplied in three
different ways:

1) A default task runner can be specified when creating the Graph
   this task runner is used for all stage/nodes that aren't otherwise
   supplied with one.
2) A specific task runner can be supplied when adding the node to the
   graph.
3) The node itself can insist on using a particular task runner.

This last method is for nodes that have specific dispatching
requirements. For example, the ffmpeg decoders need to run on a
single thread, so they create a thread and use its task runner.

This approach allows the graph builder to control what stage/node
pairs run on what threads. For now, the MultiprocTaskRunner is used
by default, but this may change in the future.

Change-Id: Id32aaabb4f8fa59ef9018330672e6998c12f0cff