Merge SingleServiceViewApp into SingleServiceApp.
We can always implement ViewProvider with a default implementation
that drops the view provider request. This is the same behavior as
not implementing ViewProvider in the first place, but written with
less boilerpate.

Since we now have a default implementation of ViewProvider.CreateView(),
the no-op implementation can be omitted in derived classes that don't
do anything else.

Collateral cleanup:

* Typos.

Change-Id: Ia1b068062cd048c983121b302ffe5a700af5b257