apps/media: add multi-stage tasks
This CL adds multi-stage tasks to the in-proc framework. These will
be used in the future to make dynamic changes to a graph. The graph
now supports a PostTask call that accepts any number of node (stage)
references. The task is run with all the specified nodes/stages
'acquired', meaning that their task queues have been suspended. Once
the task completes, the nodes/stages are released. This will allow
graphs to be edited and reconfigured with proper synchronization.

Change-Id: I6d062cc56179e401f00469082330633ceec5232f