[kernel][magenta] Add slack options to timers
The long requested feature now exposed to user mode. Now
the timer slack and be early, center or late.

The default behavior of 0 is "center" which was the existing
behavior so there no effective change on the existing
build.

Tests included with the caveat that better tests
exist already via "k timer_tests".

Change-Id: If1787e30eda4c06be003dea715e92c3f240841ca