[acpi] Standardize pci access around PIO
According to the PCIe Base Spec section 7.2 all PCIe devices
must be compatible with PCI 3.0 configuration accesses. This
change standarizes the ACPI OS abstraction layer to only use PIO
access for ACPI initialization. This results in us not needing
to do any error prone probing and cuts down on assumptions we
need to make around the hardware the system is running on.

As an added bonus, it greatly reduces the complexity of the
abstraction layer between ACPICA and magenta.

Change-Id: I3c30b73058a0c4c6c361f2d17ca8d15a27bda050