Add -A switch to build-magenta.sh to control ASan build
Without -A, the main Magenta image in out/build-magenta-HW-CPU/ is built
without ASan and a secondary image in out/build-magenta-HW-CPU-ulib/
contains just the ASan-supporting shared libraries.  The secondary image
supports any Fuchsia components selected to build with ASan.

With -A, all of Magenta userland is built with ASan and the secondary
image in out/build-magenta-HW-CPU-ulib/ contains just the non-ASan
versions of the Magenta shared libraries.  The secondary image supports
all Fuchsia components not specifically selected to build with ASan.
Using the ASan Magenta is necessary to use any Fuchsia device driver
modules built with ASan (though non-ASan drivers can still be used).

Change-Id: I6a636eedf149b376fe838706aac085b26e62dd11