[examples][ble_scanner] Add connection support
Added central-role connection support to the BLE scanner example
app. The UI now displays the connection state of peripherals and
the "Connect" button is no longer a dummy.

The application currently cannot do anything with the connection.
Once the GATT API is implemented it will be able to browse and
interact with remote services.

Change-Id: I10897d757f8737d314312f5cd7641f0ba40aa749