Map module's /data into a specific directoy in user_runner's namespace.
This means that a module's /data is not the same the root /data of the
system.
This also means that for a user, all instances of the 'same module' will see
the same /data mapped to their namespace. For example, if a user has two
email stories running - the email module in both stories will see the
same /data which is not the same as the root /data.

Currently same module just means modules that were started from the same
url. Later we may want to exapand it to a security principle.

Change-Id: I0ab3a3a92324fb5269387cabac1be2329ab834eb