[public] Add __LEAF_FN macro to <magenta/compiler.h>
This is another macro for a GCC attribute.  The more obvious and
consistent name __LEAF is not used because <magenta/compiler.h>
winds up being included in host code and GNU/Linux headers define
a macro called __LEAF.

Use __LEAF_FN in place of __attribute__((__leaf__)) in sysgen.
This eliminates the need for -Wno-unknown-attributes with Clang.

MG-1052

Change-Id: I01570d8a6ed97ca8f45b0d0e8f60c5636fc5fd81