[kernel][magenta] Fix port bug
When trying to remove the port leaker we end up creating
a kernel crash, which takes some time to repro

The issue is that CanRep() was racing with Dequeue()
when the packet is removed from the queue and the lock
is released. The effect is basically a double free
on the delete observer.

Best repro on the existing code so far is
1- apply PS1 or PS2
2- $ while USER_AUTORUN=autorun ./scripts/run-magenta-x86-64 -b -k; do :;
done
3- wait for ~5 minutes (you need kvm)

with autorun being:
msleep 50
/boot/test/sys/exception-test
msleep 50
dm poweroff

While here we
- remove duplicated fields key_ annd handle_
- add a couple more debug asserts

MG-1093 #done
MG-1094 #done

Change-Id: I9d44f070c5b38bc1a7edbce8c344a0b2edacf6b3