Unit testing on device for Dart
Tests are written in the normal Dart package:test way, a series of files
with a `main()` method that invokes `test()` and other methods from
`package:test/test.dart`. The `dart_fuchsia_test` gn template builds a
dartx executable that includes all of the tests from a test directory
and a generated main file that with the help of a helper runs them
through the Dart testing flow. That executable can be run and it prints
the resuts out.

This currently does not integrate with runtests, the CQ or waterfall
and calls all kinds of internal APIs from `package:test`, but even with
this I've found and fixed a lot of bugs in the Dart FIDL bindings in
just a few hours.

Change-Id: I6b235a9485a6b55af0c6f6d440b70b93866df0b8