[psutils] Let killall take fnmatch patterns.
We now name some Fuchsia content handlers after the names of the
applications they're running, so the `dart_content_handler` names itself
`dart:my_app.dartx` when it's running `my_app.dartx`. This makes them
harder to kill with `killall`.

This change allows the argument that's passed to `killall` to be
interpreted as an fnmatch pattern so I can do `killall dart:*` and all
the dart content handlers will be killed.

Change-Id: Ib349877e8cfda9d563f08ea3a41dd3be297edcd1