Add Amber control UI and FIDL API
Add the beginnings of a FIDL API to Amber. This first change
only implements a portion of the first set of APIs.

Add a command line utility that uses the FIDL interface.

Change the way the Amber daemon is started so that the daemon
is the thing the command line tool connects to.

Change-Id: Ia68ca71ee1d43f52de9d4eea4cc2d752e0c3f7cb