[docs] Symbolizer markup specification
The new docs/symbolizer_markup.md document provides a specification
for Fuchsia Symbolizer Markup Format.  This is a format already used
by the LLVM sanitizer runtime support for Fuchsia.  It will be
eventually be used for the dynamic linker's memory layout logging,
crashlogger and kernel backtrace printing, and other things.  There
will eventually be one or more implementations of a symbolizing
filter (as defined in the document), and specific practical
documentation on using them.

Change-Id: If9c06565bc7bcdd4f58290e32f3909158d181ff5