Shell script to run Ledger benchmarks
Currently the set benchmarks to run continuously is specified in a JSON
file, which is used to invoke tests over a TCP connection. I'm changing
the Fuchsia recipe to use the runtests command, which looks for tests to
execute in a known directory.

So there will need to be something there that executes the right things.
I'm copying the set of benchmarks to run from the existing JSON file,
although I see it's not all of them.

For background and rationale see:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/54465

Change-Id: Ifb9b6de359dace04183769289cb00adb22951973