[wlantool] Add connect command
Connect command changes the SSID to connect, which is initialized from
the config file.

If wlantstack finds a valid SSID in the config file, it automatically
starts a state transition: Scan -> Join -> Authenticate -> Associate.
But the state transition may fail for some reason. Or if there is no
SSID specified in the config file, the state transition is not started.

The user can use "connect" command to change the SSID and the scan
interval to connect to a new SSID.

Change-Id: I6c9d8c8eae5157654ed152382dedecfefca4e720