[hypervisor] Serialise uart r/w regs and handle DL
Changes the trap to only deal with writes to THR. Reads and writes to
config registers were sometimes happening out of sequence, causing weird
glitches.

Ignore writes to IER and THR when divisor latch is enabled. We don't
actually take action on requests to set the connection speed, but we
shouldn't allow them to leak into the output or interrupt registers.

Change-Id: I06923b103ea0e3d60229b249f62d34b2496c41a4