Fix LedgerRepository storage location.
The repository data should be stored in the repository content_path,
not in base_path, since only content_path is erased when the repository
is erased.

Arguably EraseRepository() should actually erase base_path, not just
content_path, in order to fully revert the effect of GetRepository(),
but that's left for a followup change if indeed desired.

LE-304 #done

Change-Id: If766c8f435f3a1e559a46f002e6e8f794dd0a2a6