Updated various paths.
The C++ changes are not ideal as it'd be better to properly set an include path as opposed to including relatively to the root of the repository.
That being said, a lot of this code will move around in the near future and at that point we can pick the right include strategy.

Change-Id: I72033bcf74f865ff88dac734a6945b15f589d870