[fs] Drastically improve fsck performance
For files which are not full of indirect / doubly indirect
blocks (i.e., most files), the process of iterating over
all possible logical blocks is extremely slow. For the resize
test, which creates 100,000 empty files, this means that fsck
iterates over all possible blocks for each file, even though
those files don't actually have any allocated blocks.

As a result, fsck used to take on the order of 30-60 minutes
to complete for this test.

Now, for "GetInodeNthBno", an fsck-specific function, the output
parameter "next_n" is returned, as a suggestion for the next
block (relative to the start of the file) to check. If "n"
is the start of an unallocated indirect block, there is no reason
to check the "n+1"th block -- we now jump to the next indirect block.

The new implementation of fsck takes approximately 0.5 seconds
to iterate over all 100,000 nodes.

MG-1100 #done

Change-Id: Ia3fd21a321ad1244bc1a335add79b8d84c3b73e5