[kernel][magenta] remove system exception port
... as separate mechanism from the code.

Once we added exception propagation upwards the job
tree it became clear that the system exception handler
is nothing more than the exception handler of the
root job.

This CL keeps the current syscall semantics so no
need to change the tests or other callers but
simplifies the exception handling code.

All tests and crashlogger is functional with
this change.

Change-Id: I794a1af6cbd99e09e828e5b7152f805d0457d179