Refactors the Cobalt client FIDL service to use ShippingManager.
Also we introduce a new FIDL interface CobaltController.
Currently our only know usage of this interface is in our test app.
Possibly there will be other uses for the interface in the future.

Also we refactor the test app so that it tests using three different
strategies:
(1) Initialize the Cobalt service in manual mode and invoke
RequestSendSoon() whenever we want to send Observations.

(2) Initialize the Cobalt service in a mode where it sends every
second and then use BlockUntilEmpty() to wait for the send and
then use NumSendAttempts() and FailedSendAttempts() to check
that the sends succeeded.

(3) Don't explicitly initialize the Cobalt service but instead take
the service from the environment. In this case since we don't have a
chance to configure the scheduling parameters we don't test the
sucess of the sends at all. We only test that we can successfully
connect to and invoke the FIDL service.

Change-Id: I350349762b48e60691d4ac21cb9ad0ff17d15a14