[minfs] Fix fsck bug which incorrectly flags direntry sizes
MinFS direntry records have two distinct notions of
size: First, they have a 'minimum required size for their name',
as calculated by DirentSize, but they also have a 'physical
record length', which must be greater than or equal to the
DirentSize name.

The "kMinfsMaxDirentSize" constant is the maximum calculated
value of the DirentSize length, but it may notably be less
than the record length. This patch fixes an issue with
fsck, where it makes the (invalid) assumption that
"kMinfsMaxDirentSize" refers to the maximum record length.

Regression test added, which triggers the fsck failure
without the fix.

Change-Id: I0d9293918428cfc4ef780657659c569812e851b2