Use an even more robust termination criterion.
After all transactions are done, first wait for the conflicts to
be resolved, then force out notifications.

Conflicts resolved is detected only after the Done() call on the
conflict provider returns. This should ensure that the following
transactions really send notifications for the resolved values,
not for the values before resolution.

Also, change interleaving of transactions to force conflicts,
but that was not a problem. It's just slight more obviously
a conflict now.

FW-287 #done

Change-Id: I1ad729e6f91b405c5e8ebb4df0f157c69a9343cb