[ulib][musl] Optimized aarch64 string routines
These are taken verbatim from Linaro's cortex-strings package.
One mechanical cosmetic change to the assembly was necessary for
it to build with Clang/LLVM's assembler because the Linaro code
was using a GAS variant of the official syntax instead of what
the ARMv8 manual says to use.  We can revert that to go back
to verbatim when https://bugs.llvm.org//show_bug.cgi?id=34488 is fixed.

MG-1111 #done

Change-Id: I186ee61b9184f247ba77a27622389eee4bde493c