Starting using the network in Cobalt CI and CQ.
I don't want for the Fuchsia CI to become red or for the
Fuchsia CQ to fail if the Cobalt server goes down.

But I *DO* want the Cobalt CI to become red and the Cobalt CQ
to fail ifthe Cobal server goes down. Thi is because anyone
who notices these things is likely to be on the Cobalt team
and therefore will understand what is going on and can respond
appropriately. It is actually really useful to have that indication
that something is wrong with the servers.

As far as I know this change only affects the cobalt_client CI
and CQ.

Change-Id: I4ab16a78bb31cc8345fa24bc1b25b4bcf28a091b